software/libbase/spisdcard: fix width of address parameter
Host address parameter types should match CPU word width, so
use `unsigned long` to be correct on both 32 and 64 bit CPUs.

+++ litex/soc/software/include/base/spisdcard.h
@@ -3,4 +3,4 @@ int spi_sdcard_read_sector(uint32_t device, uint32_t lba,uint_least8_t *buf);
 
 uint8_t spi_sdcard_goidle(void);
 uint8_t spi_sdcard_readMBR(void);
-uint8_t spi_sdcard_readFile(char *, char *, uint32_t);
+uint8_t spi_sdcard_readFile(char *, char *, unsigned long);
